objects: New object class variable: id_key_types
This is an extention of 87e37e616408b19a6f1a89c8aa6582f56425e823,
and adds a new variable, id_key_types, that contains the object types
of all "id" and "id_list" keys.

This variable will be used to check if the given ID belongs to
an object with the correct object_type (for example, we aren't going
to set quoted_message to an account's ID). It will also be useful
for tests.