api: Add function for GETting and PATCHing accounts by name
An interesting caveat of this is that adding an ID named "by-name"
might break some stuff. This should probably be noted in the protocol.

The db.get_object_by_key_value_pair function was added in an earlier
commit, but it was broken. This commit fixes that function to allow
this API endpoint to work with the db_dummy backend.