meta: remove id.py and move the assign_id function to objects.py
This avoids confusion with the built-in "id" module and removes an
unnecessary file.

Early in development, the id.py file was used for the ID assigning
function as it was planned to be used in multiple places, and was
originally intended to contain more ID-related functions. However,
most other ID-related functions are also DB-related, and as such
DB commands are enough.

As the id.assign function was only used in the objects.py file,
it was moved to that file and renamed to assign_id. Should any other
file or module require the ID assignment function, it can import the
objects file (which it might as well already do), but this is usually
unnecesary as the object creation system handles ID assignments
automatically upon object creation.