tests: overhaul of pregenerated variables + more coverage
The pytest fixture previously used for pregenerating objects has been
replaced with a class. This is to prevent problems in the account by-name
methods as they break with multiple accounts with the same names, and
the pytest fixture regenerated the objects for every function. Now a
single class generates the objects beforehand, avoiding such errors.