utils: add alchemify
This is an initial version of a tool that generates SQLAlchemy ORM
tables from drywall objects.

Why use a static tool to generate a table that then gets saved into
a file instead of dynamically generating the table on each startup?
There are a couple of reasons: stability, code clarity and easier
creation of migrations (parsing a diff should be relatively simple).

As for the mentions of "pypunctum": pypunctum is going to be the
official Python API library for punctum.im, and some of drywall's
files (notably objects.py and pings.py) are going to be moved to
this new generic library, which will allow app developers to use
up-to-date object/ping definitions and implementing other useful
functions. alchemify.py will eventually be moved to the pypunctum
repository, once it is created.