db: change to SQLAlchemy, part 1
This commit adds very basic support for SQLAlchemy. At the moment,
only object-related functions are finished.

The main blocker here is that the SQLite DBAPI doesn't support
datetime, causing tests to fail. If I understand the docs correctly,
it should convert things automatically, but it doesn't. This has
yet to be investigated.

The general plan is to remove support for other DB backends entirely
and replace db.py with what is currently db_sqlalchemy.py (which is
why db_models.py is named the way it is, and not something like
db_sqlalchemy_models.py), then change the config.json to allow for
configuring the DBAPI.